Add error message if PossibleModel cannot deserialize
Ran into an issue where the modelMapper function didn't have a case for
the object passed in and it returned undefined.  The existing error
messages weren't very descriptive or helpful in tracking down the issue.

Would like to improve these messages further -- doesn't seem ideal to
fire four console.errors for one issue, but I wanted the message to be
descriptive and show as much relevant information as possible